reset project architecture for first configuration
