+++ integration/src/main/java/io/reflectoring/jiraalerts/integration/base/NavbarHeaderPanel.java
@@ -6,23 +6,17 @@ import java.util.List;
 import org.apache.wicket.Component;
 import org.apache.wicket.markup.html.panel.Panel;
 import org.apache.wicket.model.ResourceModel;
-import org.apache.wicket.spring.injection.annot.SpringBean;
 
 import de.agilecoders.wicket.core.markup.html.bootstrap.navbar.Navbar;
 import de.agilecoders.wicket.core.markup.html.bootstrap.navbar.NavbarButton;
 import de.agilecoders.wicket.core.markup.html.bootstrap.navbar.NavbarComponents;
 
-import io.reflectoring.jiraalerts.integration.admin.ApplicationStatusService;
 import io.reflectoring.jiraalerts.integration.homepage.HomePage;
-import io.reflectoring.jiraalerts.integration.jiraconfiguration.JiraConfigurationPage;
 
 /**
  * Panel for the Header Including the Navigation.
  */
 class NavbarHeaderPanel extends Panel {
-
-	@SpringBean
-	private ApplicationStatusService applicationStatusService;
 
 	/**
 	 * Constructor for NavBarHeaderPanel.
@@ -41,13 +35,6 @@ class NavbarHeaderPanel extends Panel {
 
 		NavbarButton homeButton = new NavbarButton(HomePage.class, new ResourceModel("navbarHeader.home"));
 		navbarButtons.add(homeButton);
-
-		if (applicationStatusService.isApplicationActivated()) {
-			NavbarButton jiraConfigurationButton = new NavbarButton(JiraConfigurationPage.class,
-			        new ResourceModel("navbarHeader.jira.configuration"));
-			navbarButtons.add(jiraConfigurationButton);
-		}
-
 		navbar.addComponents(NavbarComponents.transform(Navbar.ComponentPosition.LEFT, navbarButtons.toArray(new Component[navbarButtons.size()])));
 		add(navbar);
 	}
