--- integration/src/main/java/io/reflectoring/jiraalerts/integration/jiraconfiguration/JiraConfigurationPage.java
@@ -1,13 +1,0 @@
-package io.reflectoring.jiraalerts.integration.jiraconfiguration;
-
-import io.reflectoring.jiraalerts.integration.base.BasePage;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.JiraConnectionDataPanel;
-
-/** Page for Jira-Configuration. */
-public class JiraConfigurationPage extends BasePage {
-
-	/** Constructor for JiraConfigurationPage. */
-	public JiraConfigurationPage() {
-		add(new JiraConnectionDataPanel("jiraUrlConfigurationPanel"));
-	}
-}
