+++ integration/src/main/resources/io/reflectoring/jiraalerts/integration/homepage/HomePage.html
@@ -3,9 +3,6 @@
 <body>
 <wicket:extend>
     <span wicket:id="welcomeLabel"/>
-    <div wicket:id="applicationDeactivatedContainer" class="alert alert-warning">
-        <span wicket:id="applicationDeactivatedLabel" />
-    </div>
 </wicket:extend>
 </body>
 </html>
