+++ integration/src/main/resources/io/reflectoring/jiraalerts/integration/homepage/HomePage.properties
@@ -1,2 +1,1 @@
 welcome.text=Willkommen auf JiraAlerts!
-application.deactivated=Die Applikation ist deaktiviert. Bitte kontaktieren Sie Ihren Administrator oder führen Sie das initiale Setup durch.
