+++ integration/src/main/resources/io/reflectoring/jiraalerts/integration/homepage/HomePage_en.properties
@@ -1,2 +1,1 @@
 welcome.text=Welcome to JiraAlerts!
-application.deactivated=The application is deactivated. Please call your Administrator or go forward for initial setup.
