--- integration/src/main/resources/io/reflectoring/jiraalerts/integration/jiraconfiguration/JiraConfigurationPage.html
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:wicket="http://wicket.apache.org">
-<body>
-<wicket:extend>
-    <div wicket:id="jiraUrlConfigurationPanel" />
-</wicket:extend>
-</body>
-</html>
