+++ integration/src/test/java/io/reflectoring/jiraalerts/integration/admin/firstconfiguration/FirstConfigurationPageTest.java
@@ -3,32 +3,30 @@ package io.reflectoring.jiraalerts.integration.admin.firstconfiguration;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import io.reflectoring.jiraalerts.integration.base.BasePage;
-import io.reflectoring.jiraalerts.integration.wickettests.IntegrationTestConfiguration;
 import org.apache.wicket.util.tester.WicketTester;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import io.reflectoring.jiraalerts.integration.wickettests.IntegrationTestConfiguration;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = IntegrationTestConfiguration.class)
 public class FirstConfigurationPageTest {
 
-    @Autowired
-    private WicketTester wicketTester;
+	@Autowired
+	private WicketTester wicketTester;
 
-    @Before
-    public void setUp() {
-        wicketTester.startPage(FirstConfigurationPage.class);
-    }
+	@Before
+	public void setUp() {
+		wicketTester.startPage(FirstConfigurationPage.class);
+	}
 
-    @Test
-    public void rendersSuccessfully() throws Exception {
-        wicketTester.assertRenderedPage(FirstConfigurationPage.class);
-    }
+	@Test
+	public void rendersSuccessfully() throws Exception {
+		wicketTester.assertRenderedPage(FirstConfigurationPage.class);
+	}
 }
