+++ integration/src/test/java/io/reflectoring/jiraalerts/integration/wickettests/IntegrationTestConfiguration.java
@@ -7,33 +7,15 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
 import io.reflectoring.jiraalerts.base.wickettests.TestConfiguration;
-import io.reflectoring.jiraalerts.integration.admin.ApplicationStatusService;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.JiraConnectionConfigurationService;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.JiraConnectionDataMapper;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.persistence.JiraConnectionDataRepository;
+import io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration.JiraConnectionDataRepository;
 
 @Configuration
 @Import(TestConfiguration.class)
 public class IntegrationTestConfiguration {
 
 	@Bean
-	public JiraConnectionConfigurationService jiraConnectionConfigurationService() {
-		return mock(JiraConnectionConfigurationService.class);
-	}
-
-	@Bean
 	public JiraConnectionDataRepository jiraConnectionDataRepository() {
 		return mock(JiraConnectionDataRepository.class);
-	}
-
-	@Bean
-	public ApplicationStatusService applicationStatusService() {
-		return mock(ApplicationStatusService.class);
-	}
-
-	@Bean
-	public JiraConnectionDataMapper jiraConnectionDataMapper() {
-		return mock(JiraConnectionDataMapper.class);
 	}
 
 }
