+++ jira-component/src/main/java/io/reflectoring/jiraalerts/jiracomponent/admin/firstconfiguration/JiraConnectionDataRepository.java
@@ -1,4 +1,4 @@
-package io.reflectoring.jiraalerts.jiracomponent.configuration.persistence;
+package io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
