+++ jira-component/src/main/resources/io/reflectoring/jiraalerts/jiracomponent/admin/firstconfiguration/FirstConfigurationPanel.html
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <html xmlns="http://www.w3.org/1999/xhtml" xmlns:wicket="http://wicket.apache.org">
 <body>
-<wicket:extend>
-    <div wicket:id="JiraConnectionDataPanel" />
-</wicket:extend>
+<wicket:panel>
+</wicket:panel>
 </body>
 </html>
