--- jira-component/src/main/resources/io/reflectoring/jiraalerts/jiracomponent/configuration/JiraConnectionDataPanel.html
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:wicket="http://wicket.apache.org">
-<body>
-<wicket:panel>
-
-    <form wicket:id="connectionUrlForm" class="form-group">
-        <fieldset>
-        <div wicket:id="connectionUrlPanel"/>
-        <div wicket:id="connectionUsernamePanel"/>
-        <div wicket:id="connectionPasswordPanel"/>
-        <a wicket:id="submitNewConnectionUrlLink" class="btn btn-primary">
-            <wicket:message key="savebutton.text"/>
-        </a><br><br/><br/>
-        <div wicket:id="feedback"/>
-        </fieldset>
-    </form>
-</wicket:panel>
-</body>
-</html>
