--- jira-component/src/main/resources/io/reflectoring/jiraalerts/jiracomponent/configuration/JiraConnectionDataPanel.properties
@@ -1,6 +1,0 @@
-connectionUrl.label=JIRA URL
-connectionUsername.label=JIRA Benutzername
-connectionPassword.label=JIRA Passwort
-savebutton.text=Speichern
-UrlValidator=Die Eingabe entspricht nicht dem Format einer gültigen URL.
-jiraconfig.saved=JIRA Konfiguration erfolgreich gespeichert
