--- jira-component/src/main/resources/io/reflectoring/jiraalerts/jiracomponent/configuration/JiraConnectionDataPanel_en.properties
@@ -1,6 +1,0 @@
-connectionUrl.label=JIRA URL
-connectionUsername.label=JIRA Username
-connectionPassword.label=JIRA Password
-savebutton.text=Save
-UrlValidator=Your input doesn´t match the pattern of an valid URL.
-jiraconfig.saved=Successfully saved JIRA configuration
