+++ jira-component/src/test/java/io/reflectoring/jiraalerts/jiracomponent/admin/firstconfiguration/FirstConfigurationServiceTest.java
@@ -1,0 +1,8 @@
+package io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration;
+
+import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+public class FirstConfigurationServiceTest {
+
+}
