+++ jira-component/src/test/java/io/reflectoring/jiraalerts/jiracomponent/wickettests/JiraComponentTestConfiguration.java
@@ -4,18 +4,15 @@ import static org.mockito.Mockito.mock;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
-import io.reflectoring.jiraalerts.jiracomponent.configuration.JiraConnectionConfigurationService;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.JiraConnectionDataMapper;
-import io.reflectoring.jiraalerts.jiracomponent.configuration.persistence.JiraConnectionDataRepository;
+import io.reflectoring.jiraalerts.base.wickettests.TestConfiguration;
+import io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration.JiraConnectionDataMapper;
+import io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration.JiraConnectionDataRepository;
 
 @Configuration
+@Import(TestConfiguration.class)
 public class JiraComponentTestConfiguration {
-
-	@Bean
-	public JiraConnectionConfigurationService jiraConnectionConfigurationService() {
-		return mock(JiraConnectionConfigurationService.class);
-	}
 
 	@Bean
 	public JiraConnectionDataRepository jiraConnectionDataRepository() {
