--- src/main/java/io/reflectoring/jiraalerts/application/CipherTest.java
@@ -1,4 +1,0 @@
-package io.reflectoring.jiraalerts.application;
-
-public class CipherTest {
-}
