+++ src/main/java/io/reflectoring/jiraalerts/application/JiraAlertsApplication.java
@@ -1,9 +1,7 @@
 package io.reflectoring.jiraalerts.application;
 
-import io.reflectoring.jiraalerts.integration.admin.FirstConfigurationService;
 import org.apache.wicket.Page;
 import org.apache.wicket.protocol.http.WebApplication;
-import org.apache.wicket.spring.injection.annot.SpringBean;
 import org.apache.wicket.spring.injection.annot.SpringComponentInjector;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
@@ -12,9 +10,8 @@ import org.springframework.stereotype.Component;
 import de.agilecoders.wicket.core.Bootstrap;
 import de.agilecoders.wicket.core.settings.BootstrapSettings;
 
-import io.reflectoring.jiraalerts.integration.admin.AdministrationPage;
 import io.reflectoring.jiraalerts.integration.homepage.HomePage;
-import io.reflectoring.jiraalerts.integration.jiraconfiguration.JiraConfigurationPage;
+import io.reflectoring.jiraalerts.jiracomponent.admin.firstconfiguration.FirstConfigurationService;
 
 @Component
 public class JiraAlertsApplication extends WebApplication {
@@ -34,7 +31,6 @@ public class JiraAlertsApplication extends WebApplication {
 	protected void init() {
 		super.init();
 		getComponentInstantiationListeners().add(new SpringComponentInjector(this, applicationContext));
-//		mountPages();
 
 		BootstrapSettings bootstrapSettings = new BootstrapSettings();
 		Bootstrap.install(this, bootstrapSettings);
@@ -42,12 +38,7 @@ public class JiraAlertsApplication extends WebApplication {
 		getRequestCycleListeners().add(new FirstConfigurationListener());
 	}
 
-//	private void mountPages() {
-//		mountPage("/jira-configuration", JiraConfigurationPage.class);
-//		mountPage("/administration", AdministrationPage.class);
-//	}
-
-	public boolean isFirstConfiguration(){
+	boolean isFirstConfiguration() {
 		return firstConfigurationService.isFirstConfiguration();
 	}
 }
